Add files via upload
I attempted to add an admin gui. I attempted to create a create user option. I also edited the loginController to check to see if the password is an admin's and if it is it redirects you to the admin view and from there you can select userview. I would like to be able to connect to the database to be able to test everything as soon as possible.