Add files via upload
If you have problems with it building after importing the extracted stuff in a folder of the same name:
Check the name for the paths related to the project name. I changed it to version5.0 and while i think that is fine for the zipped file,
I know it can cause problems if you just change the name of the folder you are keeping things in while you are working on it.